% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPlotCoordPix.R
\name{.determFigMargPix}
\alias{.determFigMargPix}
\title{Estimate size/distance of margin to edge of image (png) in pixels}
\usage{
.determFigMargPix(marg, res, callFrom = NULL, silent = FALSE, debug = FALSE)
}
\arguments{
\item{marg}{(numeric) distance of margin in inch (as given in \code{par(mar=c(  ))})}

\item{res}{(numeric, length=1) resolution of image (png)}

\item{callFrom}{(character) allows easier tracking of messages produced}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additonal messages for debugging}
}
\value{
This function returns a numeric vetctor matrix with (estimated) distance to figure margins in pixels
}
\description{
This function allows estimating size/distance of margin to edge of image (png) in pixel and return numeric vector.
}
\examples{
.determFigMargPix(c(5,4,4,2),100)
}
\seealso{
\code{\link{convertPlotCoordPix}}, \code{\link{mouseOverHtmlFile}}
}
