% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlotW.R
\name{.levIndex}
\alias{.levIndex}
\title{Transform levels into index}
\usage{
.levIndex(dat, asSortedLevNa = FALSE)
}
\arguments{
\item{dat}{(numeric) initial levels (main input)}

\item{asSortedLevNa}{(logical)}
}
\value{
This function retruns a color-gradient based on p-Values
}
\description{
This function transforms levels into index. This function has been depreciated, please use wrMisc::levIndex() insted.
}
\examples{
library(wrMisc)
}
\seealso{
(for PCA) \code{\link{plotPCAw}})
}
