% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlotW.R
\name{.sampNoDeMArrayLM}
\alias{.sampNoDeMArrayLM}
\title{Locate sample index from index or name of pair-wise comparisons}
\usage{
.sampNoDeMArrayLM(
  MArrayObj,
  useComp,
  groupSep = "-",
  lstMeans = "means",
  lstP = "BH",
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{MArrayObj}{(MArray type objct) main input}

\item{useComp}{(matrix) types of pair-wise comparisons to be performed}

\item{groupSep}{(character) separator used with pair-wise grouping}

\item{lstMeans}{(character) type of summarization, default is 'means'}

\item{lstP}{(character) type of multiple testing correction data to choose from \code{MArrayObj}}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) supplemental messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a integer vector of indexes
}
\description{
This function helps locating sample index from index or name of pair-wise comparisons
}
\examples{
aa <- 1:5
}
\seealso{
(for PCA) \code{\link{plotPCAw}})
}
