% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouseOverHtmlFile.R
\name{.serachColName}
\alias{.serachColName}
\title{Search Column Name	 
 
This function provides help when seraching column names}
\usage{
.serachColName(
  x,
  searchColNa,
  plusLowerCaps = TRUE,
  returnList = TRUE,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{x}{(matrix or data.frame) main input}

\item{searchColNa}{(character)}

\item{plusLowerCaps}{(logical) add lower caps to search}

\item{returnList}{(logical)}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
integer vector with index of colnames found or list with $foundNa and $remainNa
}
\description{
Search Column Name	 
 
This function provides help when seraching column names
}
\examples{
mat1 <- matrix(1:6, ncol=3, dimnames=list(NULL, LETTERS[1:3]))
.serachColName(mat1, c("C","F","A"))
}
\seealso{
\code{\link{convertPlotCoordPix}}; use \code{\link[wrMisc]{htmlSpecCharConv}} to convert special characters for proper html display
}
